+++ conf/dataset/synthetic.yaml
@@ -1,3 +1,3 @@
-name: "simple_ob_ontology"
+name: "simple"
 DB_name: "synthetic_dialogues"
 domain: "ALL"
