+++ models/dialogue_policy/supervised_learning/LedPolicy.py
@@ -11,8 +11,6 @@ class Led(EmbeddingPolicy, ABC):
     def __init__(self, config: dict, n_actions: int):
         super().__init__(config, n_actions)
 
-        print(self.num_features)
-
         self.model = nn.LSTM(input_size=self.num_features,
                              hidden_size=self.hparams.hidden_size,
                              num_layers=self.hparams.num_layers,
@@ -21,7 +19,6 @@ class Led(EmbeddingPolicy, ABC):
 
         self.dense_transformer = create_embedding_layer(
             self.hparams.hidden_size,
-            self.hparams.embedding_space,
             self.hparams.dropout_dialogue,
             self.hparams.regularization_constant
         )
