+++ service/EvaluateService.py
@@ -2,6 +2,8 @@ import pandas as pd
 import wandb
 from typing import List
 
+from tqdm import tqdm
+
 from service.Pipeline import Pipeline
 import pytorch_lightning as pl
 
@@ -19,21 +21,23 @@ class EvaluateService(Pipeline):
         entities = []
         slots = []
         pre_actions = []
+        dialogue_id = []
 
-        for record_index in indexes:
+        for record_index in tqdm(indexes, desc='Transforming embeddings'):
             intentions.append(self.dataset['Intention'][record_index])
             slots.append(self.dataset['Slots'][record_index])
             pre_actions.append(self.dataset['Prev_action'][record_index])
             entities.append(self.dataset['Entities'][record_index])
+            dialogue_id.append(self.dataset['Dialogue_ID'][record_index])
 
-        return intentions, entities, slots, pre_actions
+        return intentions, entities, slots, pre_actions, dialogue_id
 
     def _update_test_results(self, test_results: dict):
-        intentions, slots, pre_actions, dialogue_id = self._transform_binary_embeddings(test_results['Index'])
+        intentions, entities, slots, pre_actions, dialogue_id = self._transform_binary_embeddings(test_results['Index'])
         test_results['Dialogue_ID'] = dialogue_id
         test_results['Intentions'] = intentions
-        test_results['Predictions'] = test_results['Predictions']
-        test_results['Labels'] = test_results['Labels']
+        test_results['Entities'] = entities
+        test_results['Pre_actions'] = pre_actions
         test_results['Slots'] = slots
 
     def run(self, data: object = None) -> object:
